% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_achieved.R
\name{hr_achieved}
\alias{hr_achieved}
\title{Percent of students achieving healthy relationship outcome
Determine the percent of students achieving the healthy releationship outcome}
\usage{
hr_achieved(Time)
}
\arguments{
\item{Write}{in Pre or Post for time the survey was taken
hr_achieved()}
}
\description{
Percent of students achieving healthy relationship outcome
Determine the percent of students achieving the healthy releationship outcome
}

