% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_achieved.R
\name{hr_achieved_matched}
\alias{hr_achieved_matched}
\title{Percent of students achieving healthy relationship outcome
Determine the percent of students achieving the healthy releationship outcome}
\usage{
hr_achieved_matched()
}
\description{
hr_achieved_matched()
}

