% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_achieved.R
\name{hr_achieved_post}
\alias{hr_achieved_post}
\title{Percent of students achieving healthy relationship outcome 
Determine the percent of students achieving the healthy releationship outcome at Post Survey Time}
\usage{
hr_achieved_post()
}
\description{
hr_achieved_post()
}

