% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_hr_achieved.R
\name{hr_achieved_pre}
\alias{hr_achieved_pre}
\title{Percent of students achieving healthy relationship outcome 
Determine the percent of students achieving the healthy releationship outcome at Pre Survey Time outcome}
\usage{
hr_achieved_pre()
}
\description{
hr_achieved_pre()
}

